feat: :hammer: use locale date and user utils
